Upgrade injected JavaScript to signature_pad v1.5.3
Most of the time my Android devices show a blank image for the signature
page, and do not render any brush strokes. This appears to have been a
bug that was fixed in the latest version of signature_pad, so I upgraded
the injected Javascript to be 1.5.3. My android and iOS devices both now
work with 1.5.3.

The changes I made were:
1. Copy/Paste the code from
https://github.com/szimek/signature_pad/blob/v1.5.3/signature_pad.js
into the 'content' var in signaturePad.js
2. Replace 4 spaces with 2 spaces (following your convention)
3. Replace all // comments with /* */
4. Copy/Paste the comment at the top of the file