android: Intermittent signature failures due to 0 width/height
Signatures on Android occasionally fail due to document.body.clientWidth
and window.innerWidth (as well as document.body.clientHeight and
window.innerHeight) being 0 on initialization.

I attempted to fix this by adding a setTimeout with a 300ms delay, but
even that was not a 100% effective fix. This issue is most likely
related to:
http://tripleodeon.com/2011/12/first-understand-your-screen/
http://stackoverflow.com/questions/10610743/android-browsers-screen-width-screen-height-window-innerwidth-window-inner

Note that I attempted to apply the ReactNative measured dimensions
instead of using document.body.clientWidth, but that broke iOS.

This fix will not address any rendering issues related to rotating,
since we would need to register a callback and re-initialize the canvas
element on rotation.